+++ cmd/repl.go
@@ -53,6 +53,7 @@ func NewREPLCommand() *cobra.Command {
 
 	cmd.Flags().BoolP("user-only", "u", false, "Disable built-in rules")
 	cmd.Flags().Bool("no-test-inputs", false, "Disable loading test inputs")
+	cmd.Flags().SetNormalizeFunc(normalizeFlag)
 	return cmd
 }